Moved assesmentAccepted domain event creation from AssessmentService to Cas1AssessmentDomainEventService
